+++ Environment.hpp
@@ -14,7 +14,7 @@ private:
 public:
   Environment(int field_legth, int field_width);
   //Função para calcular distância entre dois nós
-  double distance(Node& current, Node& target);
+  float distance(Node& current, Node& target);
   //Função para calcular um ponto aleatório no campo
   Node randomState();
   //Checar colisão com as extremidades
@@ -24,8 +24,8 @@ public:
 class ObstacleGrid{
 public:
   //Vetor de obstáculos (raio, Ponto)
-  vector<pair<double, Node>> _grid;
-  ObstacleGrid(vector<pair<double, Node>> grid);
+  vector<pair<float, Node>> _grid;
+  ObstacleGrid(vector<pair<float, Node>> grid);
   //Verificar se ocorre colisão entre current e target
   bool checkObstacleColision(Node& current, Node& target);
 };
